+++ Magic/CHANGELOG.md
@@ -22,6 +22,7 @@
  - Add option to fix invisible players on login if you have gotten some permanently invisible players:
    `mconfig configure config allow_player_persistent_invisibility false`
  - Add ModifyAttribute action
+ - Add full_stack parameter to TakeItem action, set to false to only take 1 item from a stack
 
 # 8.5.2